+++ basics/close-account/native/ts/instructions/index.ts
@@ -1,7 +1,0 @@
-export * from "./close";
-export * from "./create";
-
-export const MyInstruction = {
-  CreateUser: 0,
-  CloseUser: 1,
-} as const;